Make it work on socketcluster + remove duplicate file.
Dispatcher did not work on socketcluster,
most likely due to that socketcluster,
unlike socket.io, does not have an `io.use` function.
Also removes duplicate file, which was a source of confusion,
(keeping/updating the reference to the one starting with uppercase,
as that seem to be the convention in the repository).